backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
